cephadm: move container engine types and funcs to container_engines.py
Pair-programmed-with: Adam King <adking@redhat.com>
